+++ src/main/java/edu.greenblitz.gblib/encoder/IEncoder.java
@@ -19,7 +19,7 @@ public interface IEncoder {
      *
      * @return raw amount of ticks so far
      */
-    int getRawTicks();
+    double getRawTicks();
 
     /**
      *
